requirements.txt
aiogram==3.4.1
sqlalchemy==2.0.27
asyncpg==0.29.0
redis==5.0.1
pydantic-settings==2.2.1
alembic==1.13.1
python-dotenv==1.0.1
